+++ src/app/gt_mainwin.cpp
@@ -218,13 +218,6 @@ GtMainWin::GtMainWin(QWidget* parent) : QMainWindow(parent),
             SLOT(openAboutModulesDialog()));
 
     loadPerspectiveSettings();
-
-    // gui logger destination
-    gt::log::Logger& logger = gt::log::Logger::instance();
-
-    auto dest = gt::log::makeSignalSlotDestination(this, &GtMainWin::onLogMessage);
-
-    logger.addDestination(GT_CLASSNAME(GtMainWin), std::move(dest));
 }
 
 GtMainWin::~GtMainWin()
@@ -1259,14 +1252,6 @@ GtMainWin::onWidgetStructureClicked()
     {
         widgetStructureHelper(wid, 1);
     }
-}
-
-void
-GtMainWin::onLogMessage(const QString& msg, int level,
-                        const GtLogDetails& details)
-{
-    GtLogErrorMessageBox::display(gt::log::levelFromInt(level),
-                                  msg, details, this);
 }
 
 void